Cleaned the project.
